Rename: #4 - dto.event.likes로 좋아요 관련 dto 이동
