+++ src/main/java/com/teamddd/duckmap/controller/LikeController.java
@@ -8,8 +8,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.teamddd.duckmap.dto.LikeRes;
 import com.teamddd.duckmap.dto.Result;
+import com.teamddd.duckmap.dto.event.likes.LikeRes;
 
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
